refactor: 💡 code cleanup after rebase
