+++ src/app/parties/(room)/[id]/page.tsx
@@ -11,7 +11,7 @@ import { PartyroomAvatars } from '@/widgets/partyroom-avatars';
 import { PartyroomChatPanel } from '@/widgets/partyroom-chat-panel';
 import { PartyroomDisplayBoard } from '@/widgets/partyroom-display-board';
 import { DjingDialog } from '@/widgets/partyroom-djing-dialog';
-import { PartyroomUserPanel } from '@/widgets/partyroom-user-panel';
+import { PartyroomUserPanel } from '@/widgets/partyroom-participant-panel';
 import { Sidebar } from '@/widgets/sidebar';
 
 const PartyroomPage = () => {
@@ -109,42 +109,5 @@ const PartyroomPage = () => {
     </>
   );
 };
-
-// function TempMembers() {
-//   const { useCurrentPartyroom } = useStores();
-//   const { members, me, currentDj } = useCurrentPartyroom((state) =>
-//     pick(state, ['members', 'me', 'currentDj'])
-//   );
-
-//   return (
-//     <div className='flexCol pt-4'>
-//       {members.map((member, index) => {
-//         const suffixValue = (() => {
-//           if (member.memberId === currentDj?.memberId) {
-//             return 'DJing';
-//           }
-//           if (member.memberId === me?.memberId) {
-//             return 'Me';
-//           }
-//           return;
-//         })();
-
-//         return (
-//           <UserListItem
-//             key={member.uid + index}
-//             userListItemConfig={{
-//               id: member.memberId,
-//               username: member.nickname,
-//               src: member.avatarFaceUri || '/images/Temp/face.png',
-//             }}
-//             menuItemList={[]}
-//             suffixType='tag'
-//             suffixValue={suffixValue}
-//           />
-//         );
-//       })}
-//     </div>
-//   );
-// }
 
 export default PartyroomPage;
