+++ src/widgets/partyroom-participant-panel/index.ts
@@ -1,0 +1,1 @@
+export { default as PartyroomParticipantPanel } from './ui/partyroom-participant-panel.component';
