+++ src/widgets/partyroom-participant-panel/ui/parts/user-grade-panel.component.tsx
@@ -6,16 +6,17 @@ import { useStores } from '@/shared/lib/store/stores.context';
 import { CollapseList } from '@/shared/ui/components/collapse-list';
 import { UserListItem } from '@/shared/ui/components/user-list-item';
 import { renderUserListItemSuffix } from '../../model/render-suffix';
-import { categorizeParticipantsByGrade } from '../../model/user-list-panel.model';
+import { categorizeParticipantsByGrade } from '../../model/user-grade-panel.model';
 
-const UserListPanel = () => {
+const UserGradePanel = () => {
   const t = useI18n();
   const participants = useFetchParticipants();
 
-  // TODO: 임시적용. 추후 dj 정보 Get, partyroom입장시 내 정보 + memberId 받아오는 api 적용 시 수정 필요. 논의 쓰레드: https://pfplay.slack.com/archives/C03Q28EAU66/p1722694732255029
-  // const { data: me } = useSuspenseFetchMe();
   const { useCurrentPartyroom } = useStores();
   const { me } = useCurrentPartyroom();
+  // const { data: djingQueue } = useFetchDjingQueue({ partyroomId }, open);
+
+  // TODO: 임시적용. 추후 dj 정보 Get, partyroom입장시 내 정보 + memberId 받아오는 api 적용 시 수정 필요. 논의 쓰레드: https://pfplay.slack.com/archives/C03Q28EAU66/p1722694732255029
   const dj = {
     memberId: 2,
   };
@@ -55,4 +56,4 @@ const UserListPanel = () => {
   );
 };
 
-export default UserListPanel;
+export default UserGradePanel;
