--- src/widgets/partyroom-user-panel/index.ts
@@ -1,1 +1,0 @@
-export { default as PartyroomUserPanel } from './ui/partyroom-user-panel.component';
